Add PreciTemp class
