+++ xclim/indicators/atmos/__init__.py
@@ -13,4 +13,5 @@ for each indicator.
 """
 from ._conversion import *
 from ._precip import *
+from ._preciptemp import *
 from ._temperature import *
